requirements.txt
PyYAML==3.11
cyrusbus==0.1.0
RPi.GPIO==0.6.3
paho-mqtt==1.1
typing==3.5.3.0
numpy==1.12.0